% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetworx.R
\name{geom_tree3}
\alias{geom_tree3}
\title{geom_tree3}
\usage{
geom_tree3(layout = "rectangular", ...)
}
\arguments{
\item{layout}{one of 'rectangular', 'slanted', 'circular', 'radial' or 'unrooted'}

\item{...}{additional parameter}
}
\value{
tree layer
}
\description{
add tree layer
}
\author{
Klaus Schliep
}
