% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetworx.R
\name{ggnetworx}
\alias{ggnetworx}
\title{ggnetworx}
\usage{
ggnetworx(tr, mapping = NULL, layout = "slanted", open.angle = 0,
  mrsd = NULL, as.Date = FALSE, yscale = "none", yscale_mapping = NULL,
  ladderize = FALSE, right = FALSE, branch.length = "branch.length",
  ndigits = NULL, ...)
}
\arguments{
\item{tr}{phylo object}

\item{mapping}{aes mapping}

\item{layout}{one of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight'}

\item{open.angle}{open angle, only for 'fan' layout}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{yscale}{y scale}

\item{yscale_mapping}{yscale mapping for category variable}

\item{ladderize}{logical}

\item{right}{logical}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{ndigits}{number of digits to round numerical annotation variable}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
drawing phylogenetic tree from phylo object
}
\examples{

require(phangorn)
data(yeast)
dm <- dist.ml(yeast)
nnet <- neighborNet(dm)
ggnetworx(nnet) + geom_tiplab2()


}
\author{
Klaus Schliep
}
